Simplify even more when bottom nav shows
